+++ context.go
@@ -197,6 +197,111 @@ func (c *Context) Query(key string) string {
 	return c.req.query(key)
 }
 
+// QueryString returns the value of a given query parameter as a string.
+func (c *Context) QueryString(key string) string {
+	return c.req.query(key)
+}
+
+// QueryInt returns the value of a given query parameter as an int.
+func (c *Context) QueryInt(key string) (int, error) {
+	str := c.req.query(key)
+	value, err := strconv.Atoi(str)
+	if err != nil {
+		return 0, err
+	}
+	return value, nil
+}
+
+// QueryUInt returns the value of a given query parameter as a uint.
+func (c *Context) QueryUInt(key string) (uint, error) {
+	str := c.req.query(key)
+	value, err := strconv.ParseUint(str, 10, 32)
+	if err != nil {
+		return 0, err
+	}
+	return uint(value), nil
+}
+
+// QueryInt8 returns the value of a given query parameter as an int8.
+func (c *Context) QueryInt8(key string) (int8, error) {
+	str := c.req.query(key)
+	value, err := strconv.ParseInt(str, 10, 8)
+	if err != nil {
+		return 0, err
+	}
+	return int8(value), nil
+}
+
+// QueryUInt8 returns the value of a given query parameter as a uint8.
+func (c *Context) QueryUInt8(key string) (uint8, error) {
+	str := c.req.query(key)
+	value, err := strconv.ParseUint(str, 10, 8)
+	if err != nil {
+		return 0, err
+	}
+	return uint8(value), nil
+}
+
+// QueryInt32 returns the value of a given query parameter as an int32.
+func (c *Context) QueryInt32(key string) (int32, error) {
+	str := c.req.query(key)
+	value, err := strconv.ParseInt(str, 10, 32)
+	if err != nil {
+		return 0, err
+	}
+	return int32(value), nil
+}
+
+// QueryUInt32 returns the value of a given query parameter as a uint32.
+func (c *Context) QueryUInt32(key string) (uint32, error) {
+	str := c.req.query(key)
+	value, err := strconv.ParseUint(str, 10, 32)
+	if err != nil {
+		return 0, err
+	}
+	return uint32(value), nil
+}
+
+// QueryInt64 returns the value of a given query parameter as an int64.
+func (c *Context) QueryInt64(key string) (int64, error) {
+	str := c.req.query(key)
+	value, err := strconv.ParseInt(str, 10, 64)
+	if err != nil {
+		return 0, err
+	}
+	return value, nil
+}
+
+// QueryUInt64 returns the value of a given query parameter as a uint64.
+func (c *Context) QueryUInt64(key string) (uint64, error) {
+	str := c.req.query(key)
+	value, err := strconv.ParseUint(str, 10, 64)
+	if err != nil {
+		return 0, err
+	}
+	return value, nil
+}
+
+// QueryFloat32 returns the value of a given query parameter as a float32.
+func (c *Context) QueryFloat32(key string) (float32, error) {
+	str := c.req.query(key)
+	value, err := strconv.ParseFloat(str, 32)
+	if err != nil {
+		return 0, err
+	}
+	return float32(value), nil
+}
+
+// QueryFloat64 returns the value of a given query parameter as a float64.
+func (c *Context) QueryFloat64(key string) (float64, error) {
+	str := c.req.query(key)
+	value, err := strconv.ParseFloat(str, 64)
+	if err != nil {
+		return 0, err
+	}
+	return value, nil
+}
+
 // Queries returns all query parameters for the req.
 func (c *Context) Queries() map[string][]string {
 	return c.req.queries()